Move StochasticDistributionClass to distribution.hpp
