+++ openscenario/openscenario_interpreter/include/openscenario_interpreter/syntax/normal_distribution.hpp
@@ -18,6 +18,7 @@
 #include <openscenario_interpreter/scope.hpp>
 #include <openscenario_interpreter/syntax/double.hpp>
 #include <openscenario_interpreter/syntax/range.hpp>
+#include <openscenario_interpreter/utility/distribution.hpp>
 
 #include <random>
 #include <string>
@@ -38,21 +39,6 @@ inline namespace syntax
  *
  * -------------------------------------------------------------------------- */
 
-template <typename DistributionT>
-struct StochasticDistributionClass
-{
-  template <typename... Ts>
-  StochasticDistributionClass(Scope & scope, Ts... xs)
-  : random_engine(scope.ref<Double>(std::string ("randomSeed")).data), distribution(xs...)
-  {
-  }
-
-  std::mt19937 random_engine;
-
-  DistributionT distribution;
-
-  auto generate() { return distribution(random_engine); }
-};
 struct NormalDistribution : public ComplexType
 {
   const Range range;
@@ -65,8 +51,6 @@ struct NormalDistribution : public ComplexType
 
   explicit NormalDistribution(const pugi::xml_node &, Scope & scope);
 
-  // TODO: implement evaluate()?
-  // Use std::normal_real_distribution from <random>
   auto evaluate() -> Object;
 };
 }  // namespace syntax
